steps: Remove old style HTTP steps
