+++ master/buildbot/newsfragments/old-style-http-steps.removal
@@ -1,0 +1,1 @@
+``GET``, ``PUT``, ``POST``, ``DELETE``, ``HEAD``, ``OPTIONS`` steps now use new-style step implementation.
